feat: error handling, move character legends to separate module
